+++ crates/cubed-application/src/ports/mod.rs
@@ -1,5 +1,6 @@
 pub mod backup_repository;
 pub mod console;
+pub mod mod_repository;
 pub mod downloader;
 pub mod file_system;
 pub mod java_manager;
@@ -9,6 +10,7 @@ pub mod resource_monitor;
 pub mod server_repository;
 
 pub use backup_repository::BackupRepository;
+pub use mod_repository::ModRepository;
 pub use console::{ConsoleLine, ConsoleCallback, ConsoleManager};
 pub use downloader::{DownloadedJar, Downloader};
 pub use file_system::FileSystemManager;